magic method that denotes whiteboxity
In the previous take on the new syntax for unified macro defs and impls,
we said `macro foo: A = e1` for blackbox macros and `macro bar <: B = e2`
for whitebox macros.

This, however, requires hacking up the syntax analyzer, which is something
that I'm proud of, but it's also something I wouldn't like to do routinely.
That's why it'd be better for us to have an alternative, compatible syntax.

One take on this is to use the same syntax that Scala 2.10 and 2.11 do
and go for the subsyntax of that syntax that is guaranteed not to be
well-typed in vanilla Scala: `def baz: C = macro { ... }`.

This notation, however, has a problem in the sense that it can't accommodate
boxity annotations. Therefore here we go for a marker method that serves
this purpose.